refactor(batcher): move InputValidationError out of batcher.ts
The batch processor needs to throw this class, and the HTTP layer branches
on `instanceof`, so there must be exactly one class identity. Importing it
from batcher.ts would close a cycle, since batcher.ts imports the
processor.

Moved to core/errors.ts and re-exported from batcher.ts, which has always
carried it in its public surface. No behaviour change.
